+++ plugin/nube/protocals/networklinker/api.go
@@ -39,10 +39,6 @@ func (inst *Instance) RegisterWebhook(basePath string, mux *gin.RouterGroup) {
 		api.ResponseHandler(body, err, ctx)
 	})
 
-	mux.PATCH(plugin.PointsWriteURL, func(ctx *gin.Context) {
-		inst.handlePointWriteProxy(ctx)
-	})
-
 	mux.DELETE(plugin.NetworksURL, func(ctx *gin.Context) {
 		body, _ := plugin.GetBODYNetwork(ctx)
 		ok, err := inst.db.DeleteNetwork(body.UUID)
@@ -53,11 +49,20 @@ func (inst *Instance) RegisterWebhook(basePath string, mux *gin.RouterGroup) {
 		ok, err := inst.db.DeleteDevice(body.UUID)
 		api.ResponseHandler(ok, err, ctx)
 	})
-	// mux.DELETE(plugin.PointsURL, func(ctx *gin.Context) {
-	//     body, _ := plugin.GetBODYPoint(ctx)
-	//     ok, err := inst.db.DeletePoint(body.UUID)
-	//     api.ResponseHandler(ok, err, ctx)
-	// })
+	mux.DELETE(plugin.PointsURL, func(ctx *gin.Context) {
+		body, _ := plugin.GetBODYPoint(ctx)
+		ok, err := inst.db.DeletePoint(body.UUID)
+		api.ResponseHandler(ok, err, ctx)
+	})
+
+	mux.PATCH(plugin.PointsURL, func(ctx *gin.Context) {
+		body, _ := plugin.GetBODYPoint(ctx)
+		point, err := inst.db.UpdatePoint(body.UUID, body, true, false)
+		api.ResponseHandler(point, err, ctx)
+	})
+	mux.PATCH(plugin.PointsWriteURL, func(ctx *gin.Context) {
+		inst.handlePointWriteProxy(ctx)
+	})
 
 	mux.GET(schemaNetwork, func(ctx *gin.Context) {
 		ctx.JSON(http.StatusOK, inst.GetNetworkSchema())